Server/webapp/requirements/base.txt
pytz===2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==7.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
#              # when upgrading django: have a look at Readme.md `Upgrading Webapp dependencies`
django==3.1.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/niwinz/django-redis
django-taggit==1.3.0

# Django Extensions
# we only use the reset_db command
django-extensions==3.0.9 # https://github.com/django-extensions/django-extensions

# Django REST Framework
djangorestframework==3.12.1  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# Translating model fields:
django-parler==2.2

# camelcase transformer
djangorestframework-camel-case==1.2.0

# dataclasses
attrs==20.2.0

# graphql
django-filter==2.4.0
graphene-django==2.13.0

# to generate short uuids
shortuuid==1.0.1

# Django Storage Adapter for media/statics for AWS S3
django-storages==1.10.1
boto3==1.15.16

# Wagtail CMS
wagtail==2.10.2
wagtailtrans==2.2.1
wagtailmenus==3.0.2

# django-webpack-loader needed for scss, js bundling of assets from static_src to static
django-webpack-loader==0.7.0

# country codes
django-countries==6.1.3

# Markdown Support
django-markupfield==2.0.0
markdown==3.3

django-dbbackup==3.3.0
